add jvmId to identify jvm instance in the logs
uses System.currentTimeMillis() at startup as the jvmId, so if two lambdas are started at the same time they will have the same lambdaId. We expect this to be rare, and in any case this can be disambiguated by using logstream name + jvmId. System.currentTimeMillis() is preferred to a UUID because it has a smaller representation, and will look distinct from the request id (which is a UUID) in the logs.